first commit
HTML script for the github pages. Made dark mode and a general chatbot UI. Will need to adjust the UI to better fit the needs of the project.

I have the Firebase function stuff in here too. Right now it does nothing. I will likely make some form of dummy code for testing in a second.